finish all pooling

+++ src/pooling.h
@@ -7,5 +7,5 @@
 void maxpooling_cpu(float*, const int, const int, float*, float*);
 void meanpooling_cpu(float*, const int, const int, float*);
 
-__global__ void maxpooling_forward(float*, const int, const int, float*, float*);
+__global__ void maxpooling_forward(float*, const int, const int, float*, int*);
 __global__ void meanpool_forward(float*, const int, const int, float*);
